move helper methods to exclusive file that they're used in
